+++ packages/relay/src/lib/services/ethService/ethCommonService/ITimestamp.ts
@@ -1,0 +1,11 @@
+// SPDX-License-Identifier: Apache-2.0
+
+export interface ITimestamp {
+  from: string;
+  to: string;
+}
+
+export interface LatestBlockNumberTimestamp {
+  blockNumber: string | null;
+  timeStampTo: string;
+}
