+++ packages/relay/tests/lib/eth/eth_getBalance.spec.ts
@@ -686,8 +686,7 @@ describe('@ethGetBalance using MirrorNode', async function () {
     });
   });
 
-  // TODO: migrate these tests to accountService.spec.ts
-  xdescribe('Calculate balance at block timestamp via getBalanceAtBlockTimestamp', async function () {
+  describe('Calculate balance at block timestamp via getBalanceAtBlockTimestamp', async function () {
     const timestamp1 = 1651550386;
 
     it('Given a blockNumber, return the account balance at that blocknumber, with transactions that debit the account balance', async () => {
@@ -696,7 +695,7 @@ describe('@ethGetBalance using MirrorNode', async function () {
         buildCryptoTransferTransaction('0.0.98', CONTRACT_ID_1, 50, { timestamp: `${timestamp1}.060890954` }),
       ];
 
-      const resultingUpdate = ethImpl.getBalanceAtBlockTimestamp(
+      const resultingUpdate = ethImpl.accountService.getBalanceAtBlockTimestamp(
         CONTRACT_ID_1,
         transactionsInBlockTimestamp,
         Number(`${timestamp1}.060890950`),
@@ -711,7 +710,7 @@ describe('@ethGetBalance using MirrorNode', async function () {
         buildCryptoTransferTransaction(CONTRACT_ID_1, '0.0.98', 50, { timestamp: `${timestamp1}.060890954` }),
       ];
 
-      const resultingUpdate = ethImpl.getBalanceAtBlockTimestamp(
+      const resultingUpdate = ethImpl.accountService.getBalanceAtBlockTimestamp(
         CONTRACT_ID_1,
         transactionsInBlockTimestamp,
         Number(`${timestamp1}.060890950`),
@@ -726,7 +725,7 @@ describe('@ethGetBalance using MirrorNode', async function () {
         buildCryptoTransferTransaction(CONTRACT_ID_1, '0.0.98', 50, { timestamp: `${timestamp1}.060890954` }),
       ];
 
-      const resultingUpdate = ethImpl.getBalanceAtBlockTimestamp(
+      const resultingUpdate = ethImpl.accountService.getBalanceAtBlockTimestamp(
         CONTRACT_ID_1,
         transactionsInBlockTimestamp,
         Number(`${timestamp1}.060890950`),
@@ -742,7 +741,7 @@ describe('@ethGetBalance using MirrorNode', async function () {
         buildCryptoTransferTransaction('0.0.98', CONTRACT_ID_1, 20, { timestamp: `${timestamp1}.060890955` }),
       ];
 
-      const resultingUpdate = ethImpl.getBalanceAtBlockTimestamp(
+      const resultingUpdate = ethImpl.accountService.getBalanceAtBlockTimestamp(
         CONTRACT_ID_1,
         transactionsInBlockTimestamp,
         Number(`${timestamp1}.060890950`),
